+++ CHANGES.rst
@@ -178,5 +178,5 @@ v1.2.0
 * Apply.status returned by API is not an "key" field anymore, but a readable string, so instead of getting "applied" and "unapplied" you might now get "Applied" and "Canceled".
 * Introduced route /projects/{project.slug}/applies/{apply.id}/ which can be PATCHED by the project owner, organization owner or organization member to modify the status of an apply by a valid key choice.
 * Apply and unapply routes are changed:
-/projects/{slug}/apply => /projects/{slug}/applies/apply
-/projects/{slug}/unapply => /projects/{slug}/applies/unapply
+/projects/{slug}/apply/ => /projects/{slug}/applies/apply/
+/projects/{slug}/unapply/ => /projects/{slug}/applies/unapply/